requirements.txt
Cython==0.25.2
Django==1.10.5
django-ckeditor==5.2.1
django-modeltranslation==0.12
django-picklefield==0.3.2
django-solo==1.1.2
Kivy==1.9.1
Kivy-Garden==0.1.4
olefile==0.44
Pillow==4.0.0
psycopg2==2.7.1
requests==2.13.0
